Expected testing, scoreboard.
The scoreboard is working as expected. It will output the name and the number of attempts of the user and if more than one player has played the game then it will add a comma after the first results and then output the results for the second user and so on.